refactor(app): extract edit mask workspace hook
